Added a comic state handler adaptor to decouple event handler [#1970]
